fix: allow creating File objects with not existing path
Previously `JsonFile` and `YamlFile` were incorrectly throwing error
during `__init__` if file was missing. Now errors are thrown only
when `read()` is called, not before that. Also, now base class `File`
is throwing this error, previously `File` was never checking if file
exists, it returned None on `read()` and was kinda inconsistent.